SA-358 Implement redux-form for entire modal
