+++ src/pages/reports/messageEvents/components/EventTypeFilters.js
@@ -1,17 +1,19 @@
 import React from 'react';
 import { Checkbox, Tooltip } from '@sparkpost/matchbox';
+import { Field } from 'redux-form';
+import { CheckboxWrapper } from '../../../../components';
 
 import styles from './AdvancedFilters.module.scss';
 
-const EventTypeFilters = ({ eventTypeDocs, checkedTypes, onChange }) => <Checkbox.Group label='Event Type'>
+const EventTypeFilters = ({ eventTypeDocs }) => <Checkbox.Group label='Event Type'>
   <div className={styles.CheckBoxGrid}>
     {eventTypeDocs.map((evtDoc) =>
-      <Checkbox
-        id={evtDoc.type}
+      <Field
+        name={evtDoc.type}
+        component={CheckboxWrapper}
         key={evtDoc.type}
-        onChange={() => onChange(evtDoc.type)}
         label={<Tooltip dark content={evtDoc.description}>{evtDoc.displayName}</Tooltip>}
-        checked={checkedTypes[evtDoc.type] || false}
+        type="checkbox"
       />
     )}
   </div>
